Minor Modifications to Frames
Ensured that the NAV_VISABILITY property of the frames is always hidden (I missed a few in the last commit), this prevents the stupid back arrow from appearing.

I also added a frame which is hidden by default but covers the entire main window, this will be used to provide a transparent help overlay which will eventually be triggered by the help button.