Added Foundational C# classes for Vertical Prototype
Added a few c# classes which will be used to formalize the information we are incorporating into the vertical prototype. This includes an event class, itinerary class, and account class.

The main window now has a dictionary of Accounts where the key is the username and the value is the account.